refactor(core): split main.rs into config/middleware/route/state; retire grandfather clause
Closes the v7 FSM pipeline cascade. PR 4c moved verb logic from main.rs
into handler.rs; this PR splits main.rs's remaining four concerns
(env parsing, axum middleware, route table, application state) into
dedicated modules so main.rs hits the 500-line ceiling that the rest
of the codebase already meets.

After this PR the grandfather clause has no further hold-out and is
deleted from AGENTS.md. Tag v7.0.0 marks "FSM pipeline complete" as
an architectural fact.

## Extractions

`core/src/config.rs` (new, 78 lines):
- `DEFAULT_MAX_WORLD_BYTES`, `DEFAULT_MAX_MEMORY_BYTES`,
  `DEFAULT_LISTEN_REPLAY_MAX`, `DEFAULT_MAX_LISTEN_CONNECTIONS`,
  `DEFAULT_COAP_MAX_IN_FLIGHT`
- `env_usize`, `env_optional_usize`, `env_nonzero_usize`
- `coap_bind_from_env`, `listen_addr`
- `should_warn_public_read`, `hmac_key_from_env_value`

`core/src/middleware.rs` (new, 74 lines):
- `add_core_response_headers` — the single Tower layer the router
  installs. Allocates `RequestId`, stashes it on extensions, stamps
  `x-request-id` / `x-elapsed-us` / `Vary: Authorization` /
  `x-content-type-options: nosniff`.
- `stamp_core_response_headers` — the helper. Lifted to `pub(crate)`
  so the white-box test in main.rs can target it directly.

`core/src/route.rs` (new, 72 lines):
- `build_app(state, max_world_bytes) -> Router` — the entire
  axum router assembly. main()'s startup just calls this.
- `world_handler` — the thin /<world> route handler (10 lines:
  OPTIONS short-circuits, every other method through pipeline::run).

`core/src/state.rs` (new, 377 lines):
- `pub(crate) struct Core` (all fields `pub(crate)` so siblings can
  read them; main.rs constructs via struct literal in main())
- `pub(crate) struct WriteOutcome`
- `pub(crate) type BoxedResponse = Box<Response>`
- All `impl Core` methods: `acquire_world_lock`, `read_world` /
  `read_world_with_etag`, `write_world` (test-only fixture),
  `world_metadata`, `delete_world_blocking`, `notify`,
  `reserve_storage` / `rollback_storage_reservation`, and
  `put_bytes` (CoAP-only entry point — HTTP goes through
  `handler::execute_put`; both retire when CoAP migrates onto the
  FSM in PR 7+).

`core/src/http_semantics.rs` (existing, +18 lines):
- `apply_meta_headers` moved here from main.rs. It's an HTTP
  semantic concern (denylist for never-persisted headers), already
  uses `is_never_persisted_header` from this file. Callers
  (handler.rs, http_semantics's own internal use, the test in
  main.rs) updated to `hs::apply_meta_headers`.

## main.rs after the cut

`core/src/main.rs` (1297 → 337 production lines):

- imports + module declarations (16 modules now)
- `pub(crate) use crate::state::*;` re-export so existing callers
  keep using `crate::Core` / `crate::WriteOutcome` without churn
  (same pattern path/pipeline/proc/response use)
- `VERSION`, `WORLD_ALLOW` consts
- `async fn main()` — env reads delegated to `config::*`, Core
  built via struct literal, router via `route::build_app`
- `print_auth_summary` (~57 lines, calls `config::should_warn_public_read`)
- `shutdown_signal` + `wait_for_shutdown_signal` (cross-platform
  helpers)
- `can_write` / `can_delete` / `can_read` / `needs_write_approve`
  / `exact_or_child` — auth policy helpers (used by handler.rs)

The `mod tests { ... }` module's existing 124 unit tests stay in
main.rs (test code is exempt from the 500-line budget per PR #106).
Test imports updated to point at the new module paths
(`crate::route::world_handler`, `crate::middleware::*`, etc).

## AGENTS.md

- **Grandfather clause section deleted entirely** (~25 lines). Both
  `core/src/main.rs` and `core/src/handler.rs` now satisfy the
  500-line ceiling; the special exemption is no longer needed.
- **Architecture Invariant for "FSM pipeline is the contract for
  new verbs" rewritten** to reflect the post-4c/4d shape:
  - Old (4c): "every new verb is one execute_* in handler.rs..."
  - New: "every new verb is one execute_* exported from
    handler.rs. Implementations live in handler.rs (dispatcher +
    light verbs) or handler/<verb>.rs (heavy verbs with their own
    audit / lock dance, e.g. delete). Use the unified primitives
    execute_read / execute_write for verbs that fit; structurally
    distinct verbs (multi-step audit, heterogeneous lock ordering)
    get their own file."
  - Names `execute_read` / `execute_write` so PR 4e/4f can
    introduce them without re-arguing the shape.
  - Leaves a documented escape hatch for DELETE-class verbs that
    don't fit the unified primitives.

## Stats

- 8 files changed: 661 insertions, 558 deletions.
- main.rs production: 1297 → 337 lines (-960, -74%).
- All four new modules well under 500 (config 78, middleware 74,
  route 72, state 377).
- handler.rs production unchanged (~395 lines, after PR #112).
- Net production: ~+103 lines (per-module imports + docs across the
  four new files).

## Test plan

- [x] `cargo test` — 124 unit tests pass (same set; test imports
      updated for new module paths).
- [x] `cargo clippy --all-targets -- -D warnings` — clean.
- [x] `cargo fmt --check` — clean.
- [x] `python sdk/tests/e2e_blackbox.py` — 233/233 checks pass.

## What's next

This is **PR 4d** — the last mandatory entry in the v7 FSM pipeline
cascade. After merge, tag `v7.0.0` to mark "FSM pipeline complete"
as an architectural fact.

Then 4e/4f (代码洁癖, optional) merge GET+HEAD and PUT+POST:
- PR 4e: `execute_read(BodyMode::{Full, HeadersOnly})` — collapses
  GET+HEAD's 160 lines to ~120, stays inside handler.rs.
- PR 4f: `execute_write(WriteMode::{Replace, Append})` — collapses
  PUT+POST, deletes handler/post.rs (folded into execute_write),
  stays inside handler.rs.

Both 4e/4f are post-v7.0.0 and produce no new files. handler.rs
ends at ~350 lines; only handler/delete.rs stays separate
(structurally distinct: dual lock + intent/commit/commit_failed
audit dance).
