+++ AGENTS.md
@@ -124,30 +124,6 @@ deepest-base first. Trying to merge a higher-up PR while a lower-down
 PR is still open creates a divergent base that GitHub will not
 auto-redirect cleanly.
 
-### Grandfather clause
-
-**Retires when `core/src/main.rs` reaches ≤ 500 production lines.**
-PR 4c brought main.rs from 1297 to 847 production lines as the
-verb implementations moved out into `handler.rs`; the post-4c
-extraction PRs (delete + post) brought handler.rs from 803 down to
-~395 production lines, so handler.rs already qualifies. main.rs is
-the sole remaining holdout.
-
-Until main.rs hits the bar, it retains the existing exemption:
-only safety fixes (P0/P1 concurrency, correctness, security) and
-extraction PRs that move code OUT into new sub-500-line modules
-may touch it. Net-new feature code MUST land in a new sub-500-line
-module.
-
-Active reduction target:
-
-- `core/src/main.rs` (~847 production) — provisional "PR 4d"
-  splits the route table + axum middleware (`add_core_response_headers`)
-  + env parsing helpers + `Core::new`/`main()` startup into
-  `route.rs` / `middleware.rs` / `config.rs`. Once that lands and
-  main.rs is ≤ 500 production lines, the grandfather clause
-  retires entirely and this section can be deleted from AGENTS.md.
-
 ### Pure-mv PRs
 
 A PR that mechanically moves N lines from one file to another counts
@@ -171,8 +147,14 @@ These are not preferences. They are the contract every change must keep.
   (validation, transactional flows, schema evolution) lives in reactors and
   SDK code, not in core. Adding policy to core is a Phoenix violation.
 - **FSM pipeline is the contract for new verbs.** Every new HTTP verb on
-  `/<world>` is one `pub(crate) async fn execute_*` in `core/src/handler.rs`
-  returning either `Phase::ExecutedRead(Response)` (read verbs),
+  `/<world>` is one `pub(crate) async fn execute_*` exported from
+  `core/src/handler.rs`. Implementations live in `handler.rs`
+  (dispatcher + light verbs) or `handler/<verb>.rs` (heavy verbs
+  with their own audit / lock dance, e.g. `delete`). Use the
+  unified primitives `execute_read` / `execute_write` for verbs
+  that fit; structurally distinct verbs (multi-step audit,
+  heterogeneous lock ordering) get their own file. Each `execute_*`
+  returns either `Phase::ExecutedRead(Response)` (read verbs),
   `Phase::CommittedWrite(Response)` (write verbs), or
   `Phase::Error { resp, reason }`. The pipeline driver
   (`pipeline::run`) handles authentication, path canonicalization,
